+++ pkg/collector/corechecks/snmp/fetch/fetch_scalar.go
@@ -1,4 +1,4 @@
-package snmp
+package fetch
 
 import (
 	"fmt"
